// src/thunder-sdk/src/types/transfer_manager/factories/TransferManagerAbi__factory.ts
/* Autogenerated file. Do not edit manually. */

/* tslint:disable */
/* eslint-disable */

/*
  Fuels version: 0.38.1
  Forc version: 0.35.5
  Fuel-Core version: 0.17.3
*/

import { Interface, Contract } from "fuels";
import type { Provider, Account, AbstractAddress } from "fuels";
import type { TransferManagerAbi, TransferManagerAbiInterface } from "../TransferManagerAbi";

const _abi = {
  types: [
    {
      typeId: 0,
      type: "()",
      components: [],
      typeParameters: null,
    },
    {
      typeId: 1,
      type: "b256",
      components: null,
      typeParameters: null,
    },
    {
      typeId: 2,
      type: "enum Identity",
      components: [
        {
          name: "Address",
          type: 5,
          typeArguments: null,
        },
        {
          name: "ContractId",
          type: 6,
          typeArguments: null,
        },
      ],
      typeParameters: null,
    },
    {
      typeId: 3,
      type: "str[21]",
      components: null,
      typeParameters: null,
    },
    {
      typeId: 4,
      type: "str[36]",
      components: null,
      typeParameters: null,
    },
    {
      typeId: 5,
      type: "struct Address",
      components: [
        {
          name: "value",
          type: 1,
          typeArguments: null,
        },
      ],
      typeParameters: null,
    },
    {
      typeId: 6,
      type: "struct ContractId",
      components: [
        {
          name: "value",
          type: 1,
          typeArguments: null,
        },
      ],
      typeParameters: null,
    },
    {
      typeId: 7,
      type: "u64",
      components: null,
      typeParameters: null,
    },
  ],
  functions: [
    {
      inputs: [],
      name: "get_exchange",
      output: {
        name: "",
        type: 6,
        typeArguments: null,
      },
      attributes: [
        {
          name: "storage",
          arguments: ["read"],
        },
      ],
    },
    {
      inputs: [
        {
          name: "exchange_contract",
          type: 6,
          typeArguments: null,
        },
      ],
      name: "initialize",
      output: {
        name: "",
        type: 0,
        typeArguments: null,
      },
      attributes: [
        {
          name: "storage",
          arguments: ["read", "write"],
        },
      ],
    },
    {
      inputs: [
        {
          name: "collection",
          type: 6,
          typeArguments: null,
        },
        {
          name: "from",
          type: 2,
          typeArguments: null,
        },
        {
          name: "to",
          type: 2,
          typeArguments: null,
        },
        {
          name: "token_id",
          type: 7,
          typeArguments: null,
        },
        {
          name: "amount",
          type: 7,
          typeArguments: null,
        },
      ],
      name: "transfer_nft",
      output: {
        name: "",
        type: 0,
        typeArguments: null,
      },
      attributes: [
        {
          name: "storage",
          arguments: ["read"],
        },
      ],
    },
  ],
  loggedTypes: [
    {
      logId: 0,
      loggedType: {
        name: "",
        type: 3,
        typeArguments: null,
      },
    },
    {
      logId: 1,
      loggedType: {
        name: "",
        type: 4,
        typeArguments: null,
      },
    },
  ],
  messagesTypes: [],
  configurables: [],
};

export class TransferManagerAbi__factory {
  static readonly abi = _abi;
  static createInterface(): TransferManagerAbiInterface {
    return new Interface(_abi) as unknown as TransferManagerAbiInterface;
  }
  static connect(id: string | AbstractAddress, accountOrProvider: Account | Provider): TransferManagerAbi {
    return new Contract(id, _abi, accountOrProvider) as unknown as TransferManagerAbi;
  }
}